Implement Definition Provider
Implement [definition provider](https://code.visualstudio.com/docs/extensionAPI/language-support#_show-definitions-of-a-symbol) to enable `Go to Definition` functionality on variables, functions, types, imports and etc.

This is not a comprehensive implementation but covers the most essential functionalities:

1. Import statement resolving both local and package dependencies (thanks to the awesome work done in completion service)
1. `is` inheritence statement resolving to either Contract or Library from direct imported files
1. `using ` statement resolving to either used Library or the target type definition
1. Annotated types (Struct, Event, Contract) for variables, function parameters, and function return parameters
1. Simple types resolve to current contract while scoped type resolves to type from specified Contract or Library
1. Call expressions resolving to either Function, Event, Struct, or Contract construction
1. Variable declreation resolving to contract storage variable or constant

All of above work for both project and package dependencies files although there might be problems if package has further external dependencies.

The major problems include:

1. Local variables are not resolved
1. Call expression can be ambigious e.g. f(x) emit Event(x) Struct(x)
1. Member access can be ambigious e.g. x.y x.f(x) x[1] x[a]

However naming ambiguity problems are not critical because the names are unlikely to be similar due to their different functionalities.