Settings enabled compiler
As opposed of having a compiler based on overriding preference, have a generic settings in which the user can select their preferred option for the workspace (remote, local, node or embedded).

Also added to the context menu the possibility for a user to compile  using a different configured compiler, local, node or remote.

This will enable in future to have specific compilers per solidity file, configured per workspace too, as we may mix and match between L2, L1 etc