Improve caching, now in mono repo we can have multiple projects cached.
Fix issue in which parsing imports will try to resolve all dependencies again, as the project will grow when re-resolving all dependencies this will cause a big performance issue.

Parsed document we have two levels of imports using the imported references. (this could be all the imported references .. but lets wait and see as normally users add the dependencies within their own contract)

On validating the document is readded to the cache, to provide a better performance